requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
scikit_image==0.18.1
tensorflow==2.5.0rc3
numpy==1.19.5
Pillow==8.2.0
skimage==0.0
gunicorn